@@ -3,8 +3,8 @@ import org.gradle.api.tasks.testing.logging.TestLogEvent
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
-    kotlin("jvm") version "1.3.72" apply false
-    id("com.google.protobuf") version "0.8.15" apply false
+    kotlin("jvm") version "1.6.20" apply false
+    id("com.google.protobuf") version "0.8.18" apply false
 
     id("io.github.gradle-nexus.publish-plugin") version "1.1.0"
 }
@@ -35,14 +35,14 @@ subprojects {
     }
 
     tasks.withType<JavaCompile> {
-        sourceCompatibility = JavaVersion.VERSION_1_7.toString()
-        targetCompatibility = JavaVersion.VERSION_1_7.toString()
+        sourceCompatibility = JavaVersion.VERSION_1_8.toString()
+        targetCompatibility = JavaVersion.VERSION_1_8.toString()
     }
 
     tasks.withType<KotlinCompile> {
         kotlinOptions {
             freeCompilerArgs = listOf("-Xjsr305=strict")
-            jvmTarget = JavaVersion.VERSION_1_6.toString()
+            jvmTarget = JavaVersion.VERSION_1_8.toString()
         }
     }
 
