@@ -285,7 +285,10 @@ object ClientCalls {
       clientCall.start(
         object : ClientCall.Listener<ResponseT>() {
           override fun onMessage(message: ResponseT) {
-            if (!responses.offer(message)) {
+            try {
+              responses.trySend(message).getOrThrow()
+            }
+            catch (e: Exception) {
               throw AssertionError("onMessage should never be called until responses is ready")
             }
           }
